// web/src/components/Appearance/interfaces.ts
export interface Tattoo {
  name: string;
  label: string;
  hashMale: string;
  hashFemale: string;
  zone: string;
  collection: string;
  opacity: number;
}

export interface TattooList {
  [key: string]: Tattoo[];
}

export interface TattoosSettings {
  items: TattooList;
  opacity: {
    min: number;
    max: number;
    factor: number;
  };
}